requirements.txt
openai==1.30.5
requests==2.32.3
pandas==2.2.2
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
